Update and Copy landnode.ks to landlongnode.ks.
I tried to use landnode.ks to land at a particular longitude every time (instead of in relation to where the sun was), but the new periapsis kept moving.
I realized the Mun was rotating whilst the ship cruised up to the node and also whilst the ship descended to the new periapsis after the node was created.
ah ha! Lines 51 through 60 make the corrections! I have left the original sun-angle code in. One could write some soft of if-then based on the parameters input so this code could do either "land before sunset" or "land at specified longitude"... p.s. thank you for all this code. It is wonderful.